refactor(clinics): move clinic photo methods inside clinic service
