+++ src/main/java/com/medspace/application/usecase/clinic/GetFilteredClinicsUseCase.java
@@ -3,7 +3,6 @@ package com.medspace.application.usecase.clinic;
 import java.util.List;
 import com.medspace.application.service.ClinicAvailabilityService;
 import com.medspace.application.service.ClinicEquipmentService;
-import com.medspace.application.service.ClinicPhotoService;
 import com.medspace.application.service.ClinicService;
 import com.medspace.domain.model.Clinic;
 import com.medspace.infrastructure.dto.clinic.ClinicQueryDTO;
@@ -19,8 +18,6 @@ public class GetFilteredClinicsUseCase {
     @Inject
     ClinicService clinicService;
     @Inject
-    ClinicPhotoService clinicPhotoService;
-    @Inject
     ClinicEquipmentService clinicEquipmentService;
     @Inject
     ClinicAvailabilityService clinicAvailabilityService;
@@ -32,7 +29,7 @@ public class GetFilteredClinicsUseCase {
 
             List<GetClinicPhotoDTO> photoDTOs =
                     queryFilterDTO.getIncludePhotos()
-                            ? clinicPhotoService.listPhotosByClinicId(clinic.getId()).stream()
+                            ? clinicService.listPhotosByClinicId(clinic.getId()).stream()
                                     .map(GetClinicPhotoDTO::new).toList()
                             : null;
