+++ src/main/java/com/medspace/application/usecase/clinicPhoto/CreateClinicPhotoUseCase.java
@@ -1,6 +1,6 @@
 package com.medspace.application.usecase.clinicPhoto;
 
-import com.medspace.application.service.ClinicPhotoService;
+import com.medspace.application.service.ClinicService;
 import com.medspace.domain.model.ClinicPhoto;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
@@ -8,9 +8,9 @@ import jakarta.inject.Inject;
 @ApplicationScoped
 public class CreateClinicPhotoUseCase {
     @Inject
-    ClinicPhotoService clinicPhotoService;
+    ClinicService clinicService;
 
     public ClinicPhoto execute(ClinicPhoto clinicPhoto) {
-        return clinicPhotoService.createPhoto(clinicPhoto);
+        return clinicService.createClinicPhoto(clinicPhoto);
     }
 }
