+++ src/main/java/com/medspace/application/usecase/clinicPhoto/DeleteClinicPhotoByIdUseCase.java
@@ -1,6 +1,5 @@
 package com.medspace.application.usecase.clinicPhoto;
 
-import com.medspace.application.service.ClinicPhotoService;
 import com.medspace.application.service.ClinicService;
 import com.medspace.domain.model.ClinicPhoto;
 import io.quarkus.security.ForbiddenException;
@@ -11,19 +10,17 @@ import jakarta.transaction.Transactional;
 @ApplicationScoped
 public class DeleteClinicPhotoByIdUseCase {
     @Inject
-    ClinicPhotoService clinicPhotoService;
-    @Inject
     ClinicService clinicService;
 
     @Transactional
     public void execute(Long photoId, Long userId) {
-        ClinicPhoto clinicPhoto = clinicPhotoService.getPhotoById(photoId);
+        ClinicPhoto clinicPhoto = clinicService.getClinicPhotoById(photoId);
         Long clinicId = clinicPhoto != null ? clinicPhoto.getClinic().getId() : null;
         Boolean isOwner = clinicService.validateClinicOwnership(clinicId, userId);
         if (!isOwner) {
             throw new ForbiddenException("Delete unauthorized");
         }
 
-        clinicPhotoService.deletePhoto(photoId);
+        clinicService.deleteClinicPhotoById(photoId);
     }
 }
