+++ src/main/java/com/medspace/application/usecase/clinicPhoto/GetClinicPhotoByIdUseCase.java
@@ -1,6 +1,6 @@
 package com.medspace.application.usecase.clinicPhoto;
 
-import com.medspace.application.service.ClinicPhotoService;
+import com.medspace.application.service.ClinicService;
 import com.medspace.domain.model.ClinicPhoto;
 import com.medspace.infrastructure.dto.clinic.GetClinicPhotoDTO;
 import jakarta.enterprise.context.ApplicationScoped;
@@ -9,10 +9,10 @@ import jakarta.inject.Inject;
 @ApplicationScoped
 public class GetClinicPhotoByIdUseCase {
     @Inject
-    ClinicPhotoService clinicPhotoService;
+    ClinicService clinicService;
 
     public GetClinicPhotoDTO execute(Long id) {
-        ClinicPhoto clinicPhoto = clinicPhotoService.getPhotoById(id);
+        ClinicPhoto clinicPhoto = clinicService.getClinicPhotoById(id);
         return new GetClinicPhotoDTO(clinicPhoto);
     }
 }
