+++ src/main/java/com/medspace/application/usecase/clinicPhoto/GetPhotosByClinicIdUseCase.java
@@ -1,6 +1,6 @@
 package com.medspace.application.usecase.clinicPhoto;
 
-import com.medspace.application.service.ClinicPhotoService;
+import com.medspace.application.service.ClinicService;
 import com.medspace.domain.model.ClinicPhoto;
 import com.medspace.infrastructure.dto.clinic.GetClinicPhotoDTO;
 import jakarta.enterprise.context.ApplicationScoped;
@@ -12,10 +12,10 @@ import java.util.List;
 @ApplicationScoped
 public class GetPhotosByClinicIdUseCase {
     @Inject
-    ClinicPhotoService clinicPhotoService;
+    ClinicService clinicService;
 
     public List<GetClinicPhotoDTO> execute(Long clinicId) {
-        List<ClinicPhoto> clinicPhotos = clinicPhotoService.listPhotosByClinicId(clinicId);
+        List<ClinicPhoto> clinicPhotos = clinicService.listPhotosByClinicId(clinicId);
         List<GetClinicPhotoDTO> clinicPhotoDTOs = new ArrayList<>();
 
         for (ClinicPhoto clinicPhoto : clinicPhotos) {
